Fix out of canvas borders

+++ libs/File2DProcessingTools/include/CSvgFileWriter.hpp
@@ -11,8 +11,8 @@ namespace File2DProcessingTools {
      */
     class CSvgFileWriter : public IVectorGraphicsDataWriter {
     private:
-        unsigned int _height = 0;
-        unsigned int _width = 0;
+        double _height = 0;
+        double _width = 0;
         double scale = 1;
         unsigned int padding_horizontal = 0;
         unsigned int padding_vertical = 0;
@@ -26,7 +26,7 @@ namespace File2DProcessingTools {
          * @param height -canvas height
          * @param width - canvas width
          */
-        CSvgFileWriter(unsigned int height, unsigned int width) noexcept;
+        CSvgFileWriter(double height, double width) noexcept;
 
         /**
          * The main method of writing points to a file
@@ -50,26 +50,26 @@ namespace File2DProcessingTools {
          * set canvas Width
          * @param height_px - canvas height
         */
-        void setCanvasHeight(unsigned int height_px) noexcept override;
+        void setCanvasHeight(double height_px) noexcept override;
 
         /**
          * set canvas Width
          * @param width_px - canvas width
         */
-        void setCanvasWidth(unsigned int width_px) noexcept override;
+        void setCanvasWidth(double width_px) noexcept override;
 
         /**
          * Set canvas size
          * @param height_px - canvas height
          * @param width_px - canvas width
          */
-        void setCanvasSize(unsigned int height_px, unsigned int width_px) noexcept override;
+        void setCanvasSize(double height_px, double width_px) noexcept override;
 
         /** Padding.first = vertical  (up-side, down-side) padding width in pixels
         * Padding.second = horizontal (left-side, right-side) padding width in pixels
         * Rescale input data to fit if needed
         */
-        bool setAlignmentCenter(std::pair<unsigned int, unsigned int> padding_px) noexcept override;
+        bool setAlignmentCenter(std::pair<double, double> padding_px) noexcept override;
 
     private:
 
@@ -83,7 +83,7 @@ namespace File2DProcessingTools {
 
         static bool isFileExists(const std::string &filepath, bool write_force);
 
-        double scaling(unsigned int height, unsigned int width) const noexcept;
+        double scaling(double  height, double  width) const noexcept;
 
         static void tryOpenFile(const std::string &filepath, std::ofstream &svg_file, bool write_force);
